refactor: do not expose ghx.Source to outside
