+++ daggerverse/gale/execution_plan.go
@@ -82,7 +82,7 @@ func (wep *WorkflowExecutionPlan) container() (*Container, error) {
 	ctr := opts.Container
 
 	// configure internal components
-	ctr = ctr.With(dag.Ghx().Source().Binary)
+	ctr = ctr.With(dag.Ghx().Binary)
 	ctr = ctr.With(dag.ActionsArtifactService().BindAsService)
 	ctr = ctr.With(dag.ActionsArtifactcacheService().BindAsService)
